+++ src/Model/Metrics/IntegerAbsoluteMetric.java
@@ -31,13 +31,13 @@ public class IntegerAbsoluteMetric implements GenericMetric {
      *
      * @author Logan MacGillivray, Ethan Morrill
      */
-    public HashMap<String, Integer> getDistance(GenericFeature feature){
-        HashMap<String, Integer> distances = new HashMap<>();
+    public HashMap<String, Double> getDistance(GenericFeature feature){
+        HashMap<String, Double> distances = new HashMap<>();
         if((feature instanceof IntegerFeature)){
             HashMap<String, GenericFeature> learnedFeature = storage.getFeature(featureName);
             Set<String> keys = learnedFeature.keySet();
             for(String key : keys) {
-                int distance = Math.abs((int)learnedFeature.get(key).getValue() - (int)feature.getValue());
+                double distance = Math.abs((int)learnedFeature.get(key).getValue() - (int)feature.getValue());
 
                 distances.put(key, distance);
             }
